Added build fat package
This can be used as a build script in most Xcode project's to build both simulator and device when pushing the build button in Xcode and automatically combining them in a universal folder, which can then be distributed. This covers the areas Apple does not support automatically in distribution of frameworks.